+++ breaking-front/src/App.js
@@ -16,7 +16,7 @@ import Profile from 'pages/Profile/Profile';
 import UserInformationProvider from 'providers/UserInformationProvider';
 import ProfileEdit from 'pages/ProfileEdit/ProfileEdit';
 import MainFeed from 'pages/MainFeed/MainFeed';
-import Post from 'pages/Post/Post';
+import PostWrite from 'pages/PostWrite/PostWrite';
 
 function App() {
   const queryClient = new QueryClient({
@@ -52,7 +52,7 @@ function App() {
                     path={PAGE_PATH.PROFILE_EDIT}
                     element={<ProfileEdit />}
                   />
-                  <Route path={PAGE_PATH.POST} element={<Post />} />
+                  <Route path={PAGE_PATH.POST} element={<PostWrite />} />
                 </Routes>
               </Layout>
             </UserInformationProvider>
